fix: resolve duplicate exports and type errors in supabase modules
